requirements.txt
Flask==2.2.2
werkzeug==2.2.2
psycopg2==2.9.5
python-dotenv==0.21.0
gunicorn==20.1.0